Refactor: Improve type safety and code structure in markdown and rehype utilities
