+++ src/libs/markdown/mdxOptions.ts
@@ -10,7 +10,6 @@ import rehypeImageSize from "../rehype/rehype-image-size";
 import rehypeKatex from "rehype-katex";
 import rehypeSlug from "rehype-slug";
 import rehypeExternalLinks from "rehype-external-links";
-import rehypeTocCustom from "../rehype/rehype-toc-custom";
 import { SsgImage } from "@/components/SsgImage";
 import { GitHubCodeLink } from "@/components/Mdx/GitHubCodeLink";
 import { Message } from "@/components/Mdx/Message";
@@ -32,9 +31,8 @@ export const defaultRemarkPlugins: PluggableList = [
 export const defaultRehypePlugins = (...paths: string[]): PluggableList => [
   rehypeSlug,
   [rehypeAutolinkHeadings, { behavior: "wrap" }],
-  rehypeTocCustom,
   [rehypeExternalLinks, { target: "_blank", rel: ["noopener", "noreferrer"] }],
-  rehypeImageSize(...paths),
+  rehypeImageSize(...paths) as any,
   [
     rehypeKatex,
     {
